feat: Add IPv4 field to External type
The code changes add an IPv4 field to the External type in the schema.graphql file. This allows for storing and retrieving the IPv4 address of an external host. The change also includes updates to the Traffic.svelte file to display the IPv4 address along with the host name in the UI.