New breadcrumbs (#200)
* New solution for breadcrumbs and titles

Pages and layouts returns meta with title and/or breadcrumbs from their
load functions.
They are combined in addPageMeta, and used by the PageHeader component

* Add other pages

* Cleanup old feature

Removed old component and features, the favourite system depends on a
few functions, so I've kept them in the component.

Had to upgrade storybook to support the featureset required by the new
PageHeader component.